+++ pkg/deployment/resources/gateway/gateway_config.go
@@ -23,16 +23,23 @@ package gateway
 import (
 	"fmt"
 	"sort"
+	"time"
 
 	bootstrapAPI "github.com/envoyproxy/go-control-plane/envoy/config/bootstrap/v3"
 	clusterAPI "github.com/envoyproxy/go-control-plane/envoy/config/cluster/v3"
 	coreAPI "github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
+	endpointAPI "github.com/envoyproxy/go-control-plane/envoy/config/endpoint/v3"
 	listenerAPI "github.com/envoyproxy/go-control-plane/envoy/config/listener/v3"
 	routeAPI "github.com/envoyproxy/go-control-plane/envoy/config/route/v3"
+	httpFilterAuthzApi "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/ext_authz/v3"
 	routerAPI "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/router/v3"
+	tlsInspectorApi "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/listener/tls_inspector/v3"
 	httpConnectionManagerAPI "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/network/http_connection_manager/v3"
+	upstreamHttpApi "github.com/envoyproxy/go-control-plane/envoy/extensions/upstreams/http/v3"
+	"github.com/golang/protobuf/ptypes/any"
 	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/types/known/anypb"
+	"google.golang.org/protobuf/types/known/durationpb"
 	"sigs.k8s.io/yaml"
 
 	shared "github.com/arangodb/kube-arangodb/pkg/apis/shared"
@@ -46,12 +53,18 @@ type Config struct {
 	Destinations ConfigDestinations `json:"destinations,omitempty"`
 
 	DefaultTLS *ConfigTLS `json:"defaultTLS,omitempty"`
+
+	IntegrationSidecar *ConfigDestinationTarget `json:"integrationSidecar,omitempty"`
+
+	SNI ConfigSNIList `json:"sni,omitempty"`
 }
 
 func (c Config) Validate() error {
 	return errors.Errors(
 		shared.PrefixResourceErrors("defaultDestination", c.DefaultDestination.Validate()),
+		shared.PrefixResourceErrors("integrationSidecar", c.IntegrationSidecar.Validate()),
 		shared.PrefixResourceErrors("destinations", c.Destinations.Validate()),
+		shared.PrefixResourceErrors("sni", c.SNI.Validate()),
 	)
 }
 
@@ -116,6 +129,41 @@ func (c Config) RenderClusters() ([]*clusterAPI.Cluster, error) {
 		def,
 	}
 
+	if i := c.IntegrationSidecar; i != nil {
+		hpo, err := anypb.New(&upstreamHttpApi.HttpProtocolOptions{
+			UpstreamProtocolOptions: &upstreamHttpApi.HttpProtocolOptions_ExplicitHttpConfig_{
+				ExplicitHttpConfig: &upstreamHttpApi.HttpProtocolOptions_ExplicitHttpConfig{
+					ProtocolConfig: &upstreamHttpApi.HttpProtocolOptions_ExplicitHttpConfig_Http2ProtocolOptions{
+						Http2ProtocolOptions: &coreAPI.Http2ProtocolOptions{},
+					},
+				},
+			},
+		})
+		if err != nil {
+			return nil, err
+		}
+		cluster := &clusterAPI.Cluster{
+			Name:           "integration_sidecar",
+			ConnectTimeout: durationpb.New(time.Second),
+			LbPolicy:       clusterAPI.Cluster_ROUND_ROBIN,
+			LoadAssignment: &endpointAPI.ClusterLoadAssignment{
+				ClusterName: "integration_sidecar",
+				Endpoints: []*endpointAPI.LocalityLbEndpoints{
+					{
+						LbEndpoints: []*endpointAPI.LbEndpoint{
+							i.RenderEndpoint(),
+						},
+					},
+				},
+			},
+			TypedExtensionProtocolOptions: map[string]*any.Any{
+				"envoy.extensions.upstreams.http.v3.HttpProtocolOptions": hpo,
+			},
+		}
+
+		clusters = append(clusters, cluster)
+	}
+
 	for k, v := range c.Destinations {
 		name := fmt.Sprintf("cluster_%s", util.SHA256FromString(k))
 		c, err := v.RenderCluster(name)
@@ -159,6 +207,33 @@ func (c Config) RenderRoutes() ([]*routeAPI.Route, error) {
 	return routes, nil
 }
 
+func (c Config) RenderIntegrationSidecarFilter() (*httpConnectionManagerAPI.HttpFilter, error) {
+	e, err := anypb.New(&httpFilterAuthzApi.ExtAuthz{
+		Services: &httpFilterAuthzApi.ExtAuthz_GrpcService{
+			GrpcService: &coreAPI.GrpcService{
+				TargetSpecifier: &coreAPI.GrpcService_EnvoyGrpc_{
+					EnvoyGrpc: &coreAPI.GrpcService_EnvoyGrpc{
+						ClusterName: "integration_sidecar",
+					},
+				},
+				Timeout: durationpb.New(500 * time.Millisecond),
+			},
+		},
+		IncludePeerCertificate: true,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	return &httpConnectionManagerAPI.HttpFilter{
+		Name: IntegrationSidecarFilterName,
+		ConfigType: &httpConnectionManagerAPI.HttpFilter_TypedConfig{
+			TypedConfig: e,
+		},
+		IsOptional: false,
+	}, nil
+}
+
 func (c Config) RenderFilters() ([]*listenerAPI.Filter, error) {
 	httpFilterConfigType, err := anypb.New(&routerAPI.Router{})
 	if err != nil {
@@ -168,6 +243,16 @@ func (c Config) RenderFilters() ([]*listenerAPI.Filter, error) {
 	routes, err := c.RenderRoutes()
 	if err != nil {
 		return nil, errors.Wrapf(err, "Unable to render routes")
+	}
+
+	var httpFilters []*httpConnectionManagerAPI.HttpFilter
+
+	if i := c.IntegrationSidecar; i != nil {
+		q, err := c.RenderIntegrationSidecarFilter()
+		if err != nil {
+			return nil, err
+		}
+		httpFilters = append(httpFilters, q)
 	}
 
 	filterConfigType, err := anypb.New(&httpConnectionManagerAPI.HttpConnectionManager{
@@ -185,14 +270,13 @@ func (c Config) RenderFilters() ([]*listenerAPI.Filter, error) {
 				},
 			},
 		},
-		HttpFilters: []*httpConnectionManagerAPI.HttpFilter{
-			{
-				Name: "envoy.filters.http.routerAPI",
-				ConfigType: &httpConnectionManagerAPI.HttpFilter_TypedConfig{
-					TypedConfig: httpFilterConfigType,
-				},
+		HttpFilters: append(httpFilters, &httpConnectionManagerAPI.HttpFilter{
+			Name: "envoy.filters.http.routerAPI",
+			ConfigType: &httpConnectionManagerAPI.HttpFilter_TypedConfig{
+				TypedConfig: httpFilterConfigType,
 			},
 		},
+		),
 	})
 	if err != nil {
 		return nil, errors.Wrapf(err, "Unable to render http connection manager")
@@ -211,7 +295,7 @@ func (c Config) RenderFilters() ([]*listenerAPI.Filter, error) {
 func (c Config) RenderDefaultFilterChain() (*listenerAPI.FilterChain, error) {
 	filters, err := c.RenderFilters()
 	if err != nil {
-		return nil, errors.Wrapf(err, "Unable to render filters")
+		return nil, err
 	}
 
 	ret := &listenerAPI.FilterChain{
@@ -228,7 +312,16 @@ func (c Config) RenderDefaultFilterChain() (*listenerAPI.FilterChain, error) {
 }
 
 func (c Config) RenderSecondaryFilterChains() ([]*listenerAPI.FilterChain, error) {
-	return nil, nil
+	if len(c.SNI) == 0 {
+		return nil, nil
+	}
+
+	filters, err := c.RenderFilters()
+	if err != nil {
+		return nil, err
+	}
+
+	return c.SNI.RenderFilterChain(filters)
 }
 
 func (c Config) RenderListener() (*listenerAPI.Listener, error) {
@@ -242,6 +335,22 @@ func (c Config) RenderListener() (*listenerAPI.Listener, error) {
 		return nil, errors.Wrapf(err, "Unable to render default filter")
 	}
 
+	var listenerFilters []*listenerAPI.ListenerFilter
+
+	if c.DefaultTLS != nil {
+		w, err := anypb.New(&tlsInspectorApi.TlsInspector{})
+		if err != nil {
+			return nil, errors.Wrapf(err, "Unable to render TLS Inspector")
+		}
+
+		listenerFilters = append(listenerFilters, &listenerAPI.ListenerFilter{
+			Name: "envoy.filters.listener.tls_inspector",
+			ConfigType: &listenerAPI.ListenerFilter_TypedConfig{
+				TypedConfig: w,
+			},
+		})
+	}
+
 	return &listenerAPI.Listener{
 		Name: "default",
 		Address: &coreAPI.Address{
@@ -252,8 +361,8 @@ func (c Config) RenderListener() (*listenerAPI.Listener, error) {
 				},
 			},
 		},
-		FilterChains: filterChains,
-
+		FilterChains:       filterChains,
+		ListenerFilters:    listenerFilters,
 		DefaultFilterChain: defaultFilterChain,
 	}, nil
 }