+++ pkg/util/tests/cast.go
@@ -18,30 +18,16 @@
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
 
-package sidecar
+package tests
 
 import (
-	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil"
+	"testing"
+
+	"github.com/stretchr/testify/require"
 )
 
-type IntegrationAuthorizationV0 struct {
-	Core *Core
-}
-
-func (i IntegrationAuthorizationV0) Name() []string {
-	return []string{"AUTHORIZATION", "V0"}
-}
-
-func (i IntegrationAuthorizationV0) Validate() error {
-	return nil
-}
-
-func (i IntegrationAuthorizationV0) Args() (k8sutil.OptionPairs, error) {
-	options := k8sutil.CreateOptionPairs()
-
-	options.Add("--integration.authorization.v0", true)
-
-	options.Merge(i.Core.Args(i))
-
-	return options, nil
+func CastAs[A any](t *testing.T, in any) A {
+	v, ok := in.(A)
+	require.True(t, ok)
+	return v
 }