+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - (Bugfix) Fix race condition in ArangoBackup
 - (Feature) Improve Gateway Config gen
 - (Feature) Integration Service TLS
+- (Feature) (Gateway) SNI and Authz support
 
 ## [1.2.42](https://github.com/arangodb/kube-arangodb/tree/1.2.42) (2024-07-23)
 - (Maintenance) Go 1.22.4 & Kubernetes 1.29.6 libraries